chore: minor refactor so that the biobankexplorer is the view and the card container is the component as it should be
